% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test.R
\name{dice}
\alias{dice}
\title{Sorensen-Dice overlap between two node sets}
\usage{
dice(a, b)
}
\arguments{
\item{a, b}{Integer vectors of node indices.}
}
\value{
A number in [0, 1]; 1 means the two sets are identical and 0 that
they are disjoint. Returns 0 when either set is empty.
}
\description{
Sorensen-Dice overlap between two node sets
}
\examples{
dice(1:20, 11:30)

}
